+++ packages/theme-patternfly-org/scripts/cli/build.js
@@ -1,6 +1,10 @@
+const path = require('path');
+const { fork } = require('child_process');
 const webpack = require('webpack');
+const { generate } = require('./generate');
+const { getConfig } = require('./helpers');
 
-async function build(webpackConfig) {
+async function buildWebpack(webpackConfig) {
   let compiler;
   try {
     compiler = webpack(webpackConfig);
@@ -39,6 +43,49 @@ async function build(webpackConfig) {
   );
 }
 
+async function execFile(file, args) {
+  const start = new Date();
+  return new Promise((res, rej) => {
+    const child = fork(path.join(__dirname, file), args);
+    function errorHandler(err) {
+      console.error(err);
+      rej();
+    };
+    function successHandler(code) {
+      if (code === 0) {
+        const duration = new Date() - start;
+        console.log('%s took %ss', file, duration / 1000);
+        res();
+      }
+      else {
+        console.error('Exited with', code);
+        rej();
+      }
+    }
+    child.on('error', errorHandler);
+    child.on('exit', successHandler);
+  });
+}
+
+async function build(cmd, options) {
+  generate(options);
+  const toBuild = cmd === 'all'
+    ? ['server', 'client']
+    : cmd;
+  const config = getConfig(options);
+  config.analyze = options.analyze;
+  // console.log('build', cmd, options.parent.cssconfig);
+  if (toBuild.includes('server')) {
+    // Need to fork since first webpack build puts pressure on GC
+    // Otherwise CircleCI will fail with 4GB RAM
+    await execFile('buildServer.js', [JSON.stringify(config)]);
+  }
+  if (toBuild.includes('client')) {
+    await execFile('buildClient.js', [JSON.stringify(config)]);
+  }
+}
+
 module.exports = {
-  build
+  build,
+  buildWebpack
 };