+++ packages/v4/CHANGELOG.md
@@ -3,6 +3,113 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.19](https://gh-redallen/patternfly/patternfly-org/compare/patternfly-org-4@4.2.17...patternfly-org-4@4.2.19) (2020-09-17)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.18](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.17...patternfly-org-4@4.2.18) (2020-09-16)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.17](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.16...patternfly-org-4@4.2.17) (2020-09-16)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.16](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.15...patternfly-org-4@4.2.16) (2020-09-15)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.15](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.14...patternfly-org-4@4.2.15) (2020-09-15)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.14](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.13...patternfly-org-4@4.2.14) (2020-09-15)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.13](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.12...patternfly-org-4@4.2.13) (2020-09-15)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.12](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.11...patternfly-org-4@4.2.12) (2020-09-15)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.11](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.10...patternfly-org-4@4.2.11) (2020-09-15)
+
+
+### Bug Fixes
+
+* **icons:** hide AnsibeTowerIcon, show AnsibleTowerIcon ([#2110](https://github.com/patternfly/patternfly-org/issues/2110)) ([9c6dcc2](https://github.com/patternfly/patternfly-org/commit/9c6dcc267465bb056da72dc69d3e5ee33d890e1a))
+
+
+
+
+
+## [4.2.10](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.9...patternfly-org-4@4.2.10) (2020-09-14)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.9](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.8...patternfly-org-4@4.2.9) (2020-09-14)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.8](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.7...patternfly-org-4@4.2.8) (2020-09-14)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
+## [4.2.7](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.6...patternfly-org-4@4.2.7) (2020-09-14)
+
+**Note:** Version bump only for package patternfly-org-4
+
+
+
+
+
 ## [4.2.6](https://github.com/patternfly/patternfly-org/compare/patternfly-org-4@4.2.5...patternfly-org-4@4.2.6) (2020-09-14)
 
 **Note:** Version bump only for package patternfly-org-4